Fix warning.
(and wondering why I depend on Travis…)

+++ Classes/GTDiffFile.m
@@ -32,7 +32,7 @@
 }
 
 - (NSString *)debugDescription {
-	return [NSString stringWithFormat:@"%@ path: %@, size: %ld, mode: %u, flags: %u", super.debugDescription, self.path, self.size, self.mode, self.flags];
+	return [NSString stringWithFormat:@"%@ path: %@, size: %ld, mode: %u, flags: %u", super.debugDescription, self.path, (unsigned long)self.size, self.mode, self.flags];
 }
 
 - (GTOID *)OID {
